fix bug in normalised tree distance calculation
and add an unused function for doing one-way normalisation. The idea
here is that a normalised tree distance might LOOK high just because
one of the trees is off at the edge of treespace somewhere. But when I
think about treespace I wonder if it’s really like that - seems like it
should be some kind of more interesting topological structure like a
donut.